+++ tests/http.test.ts
@@ -148,6 +148,12 @@ describe('HTTP tests', () => {
         assert.strictEqual(res.status, 400);
     });
 
+    it('does not create a room with invalid name: contains period', async () => {
+        const res = await axios.get(`${serverUrl}/create/index.php`)
+            .catch((error: unknown) => (error as AxiosError).response as AxiosResponse);
+        assert.strictEqual(res.status, 400);
+    });
+
     it('does not create a room with invalid name: contains space', async () => {
         const res = await axios.get(`${serverUrl}/create/foo bar`)
             .catch((error: unknown) => (error as AxiosError).response as AxiosResponse);
@@ -156,6 +162,12 @@ describe('HTTP tests', () => {
 
     it('does not join a room with invalid name: too long', async () => {
         const res = await axios.get(`${serverUrl}/012345678901234567890123456789123`)
+            .catch((error: unknown) => (error as AxiosError).response as AxiosResponse);
+        assert.strictEqual(res.status, 400);
+    });
+
+    it('does not join a room with invalid name: contains period', async () => {
+        const res = await axios.get(`${serverUrl}/index.php`)
             .catch((error: unknown) => (error as AxiosError).response as AxiosResponse);
         assert.strictEqual(res.status, 400);
     });